tools/big_five/requirements.txt
torch
torchvision
matplotlib
numpy==1.23.5
opencv-python # cv2
imageio
imageio-ffmpeg
pyyaml # yaml
tqdm
argparse
cython
scikit-image # skimage
scipy
onnxruntime
gradio
onnx
tensorflow
sklearn
shapely